requirements.txt
flake8==6.0.0
flake8-docstrings==1.7.0
aiogram==2.25.1
python-dotenv==1.0.0
pydantic==1.10.7
